+++ tests/EfRepositoryTests/Update.cs
@@ -1,14 +1,21 @@
 ﻿using GaEpd.AppLibrary.Domain.Repositories;
+using GaEpd.AppLibrary.Tests.EntityHelpers;
 using GaEpd.AppLibrary.Tests.RepositoryHelpers;
 
-namespace GaEpd.AppLibrary.Tests.LocalRepositoryTests;
+namespace GaEpd.AppLibrary.Tests.EfRepositoryTests;
 
 public class Update
 {
-    private LocalRepository _repository = default!;
+    private EfRepositoryTestHelper _helper = default!;
+
+    private EfRepository _repository = default!;
 
     [SetUp]
-    public void SetUp() => _repository = LocalRepositoryTestHelper.GetTestRepository();
+    public void SetUp()
+    {
+        _helper = EfRepositoryTestHelper.CreateRepositoryHelper();
+        _repository = _helper.GetEfRepository();
+    }
 
     [TearDown]
     public void TearDown() => _repository.Dispose();
@@ -16,17 +23,19 @@ public class Update
     [Test]
     public async Task UpdateAsync_UpdateExistingItem_ShouldReflectChanges()
     {
-        var originalEntity = _repository.Items.First();
+        var originalEntity = _repository.Context.Set<TestEntity>().First();
+        _helper.ClearChangeTracker();
         var newEntityWithSameId = new TestEntity { Id = originalEntity.Id, Name = "Xyz" };
 
         await _repository.UpdateAsync(newEntityWithSameId);
 
+        _helper.ClearChangeTracker();
         var result = await _repository.GetAsync(newEntityWithSameId.Id);
 
         using (new AssertionScope())
         {
             result.Should().BeEquivalentTo(newEntityWithSameId);
-            _repository.Items.Contains(originalEntity).Should().BeFalse();
+            _repository.Context.Set<TestEntity>().ToList().Contains(originalEntity).Should().BeFalse();
         }
     }
