+++ tests/LocalRepositoryTests/Count.cs
