+++ tests/LocalRepositoryTests/Delete.cs
