+++ tests/LocalRepositoryTests/Exists.cs
