+++ tests/LocalRepositoryTests/Find.cs
