+++ tests/LocalRepositoryTests/Insert.cs
@@ -1,4 +1,5 @@
-﻿using GaEpd.AppLibrary.Tests.RepositoryHelpers;
+﻿using GaEpd.AppLibrary.Tests.EntityHelpers;
+using GaEpd.AppLibrary.Tests.RepositoryHelpers;
 
 namespace GaEpd.AppLibrary.Tests.LocalRepositoryTests;
