+++ tests/LocalRepositoryTests/Update.cs
@@ -1,20 +1,15 @@
 ﻿using GaEpd.AppLibrary.Domain.Repositories;
+using GaEpd.AppLibrary.Tests.EntityHelpers;
 using GaEpd.AppLibrary.Tests.RepositoryHelpers;
 
-namespace GaEpd.AppLibrary.Tests.EfRepositoryTests;
+namespace GaEpd.AppLibrary.Tests.LocalRepositoryTests;
 
 public class Update
 {
-    private EfRepositoryTestHelper _helper = default!;
-
-    private EfRepository _repository = default!;
+    private LocalRepository _repository = default!;
 
     [SetUp]
-    public void SetUp()
-    {
-        _helper = EfRepositoryTestHelper.CreateRepositoryHelper();
-        _repository = _helper.GetEfRepository();
-    }
+    public void SetUp() => _repository = LocalRepositoryTestHelper.GetTestRepository();
 
     [TearDown]
     public void TearDown() => _repository.Dispose();
@@ -22,19 +17,17 @@ public class Update
     [Test]
     public async Task UpdateAsync_UpdateExistingItem_ShouldReflectChanges()
     {
-        var originalEntity = _repository.Context.Set<TestEntity>().First();
-        _helper.ClearChangeTracker();
+        var originalEntity = _repository.Items.First();
         var newEntityWithSameId = new TestEntity { Id = originalEntity.Id, Name = "Xyz" };
 
         await _repository.UpdateAsync(newEntityWithSameId);
 
-        _helper.ClearChangeTracker();
         var result = await _repository.GetAsync(newEntityWithSameId.Id);
 
         using (new AssertionScope())
         {
             result.Should().BeEquivalentTo(newEntityWithSameId);
-            _repository.Context.Set<TestEntity>().ToList().Contains(originalEntity).Should().BeFalse();
+            _repository.Items.Contains(originalEntity).Should().BeFalse();
         }
     }
