+++ tests/RepositoryHelpers/EfRepositoryTestHelper.cs
@@ -1,4 +1,5 @@
 ﻿using GaEpd.AppLibrary.Domain.Repositories.EFRepository;
+using GaEpd.AppLibrary.Tests.EntityHelpers;
 using Microsoft.EntityFrameworkCore;
 using System.Runtime.CompilerServices;
 using TestSupport.EfHelpers;
